refactor: refactor main into lib and main
